// products.d.ts
export declare const lifxProducts: {
    vid: number;
    name: string;
    products: {
        pid: number;
        name: string;
        features: {
            color: boolean;
            infrared: boolean;
            multizone: boolean;
            chain: boolean;
        };
    }[];
}[];
//# sourceMappingURL=products.d.ts.map